Fixed issue with anyone being able to rename, set type, or delete warps they don't own.
Did this by adding two admin based equivalent permissions:  niftywarp.admin.delete, niftywarp.admin.rename, niftywarp.admin.settype

You must have this admin version of the permission in order to modify/delete a warp someone else owns.

In addition, I removed the AdminCommand for now as I'm not sure when I'll decide to implement it.  But whatever time that is, it won't be before v1.0 is released